feat: created user repository
