+++ src/main/java/br/com/unimil/estilovip/domain/enumerated/TipoUsuario.java
@@ -1,0 +1,5 @@
+package br.com.unimil.estilovip.domain.enumerated;
+
+public enum TipoUsuario {
+    CLIENTE, ADMIN
+}
