+++ src/main/java/br/com/unimil/estilovip/domain/model/Endereco.java
@@ -1,0 +1,34 @@
+package br.com.unimil.estilovip.domain.model;
+
+import jakarta.persistence.Embeddable;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Endereco {
+    @NotBlank
+    private String logadouro;
+    @NotBlank
+    private String numero;
+    @NotBlank
+    private String bairro;
+    @NotBlank
+    private String cidade;
+    @NotBlank
+    private String cep;
+    private String complemento;
+
+    public Endereco(Endereco endereco) {
+        this.logadouro = endereco.getLogadouro();
+        this.numero = endereco.getNumero();
+        this.bairro = endereco.getBairro();
+        this.cidade = endereco.getCidade();
+        this.cep = endereco.getCep();
+        this.complemento = endereco.getComplemento();
+    }
+}
