--- src/main/java/br/com/unimil/estilovip/enumerated/TipoUsuario.java
@@ -1,5 +1,0 @@
-package br.com.unimil.estilovip.enumerated;
-
-public enum TipoUsuario {
-    CLIENTE, ADMIN
-}
