--- src/main/java/br/com/unimil/estilovip/model/Endereco.java
@@ -1,13 +1,0 @@
-package br.com.unimil.estilovip.model;
-
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public class Endereco {
-    private String logadouro;
-    private String numero;
-    private String bairro;
-    private String cidade;
-    private String cep;
-    private String complemento;
-}
